[feat:global] zsh install/config enhancements

+++ _scripts/core/zsh-setup.sh
@@ -38,11 +38,13 @@ fi
 
 stow_package "zsh" "" "$CURRENT_CONFIG_DIR/zsh"
 
-# Enable custom config
-if -f "$HOME/.zshrc"; then
-  if ! grep -q "source $CURRENT_CONFIG_DIR/zsh/config.zsh" "$HOME/.zshrc"; then
-    (echo; echo "source $CURRENT_CONFIG_DIR/zsh/config.zsh") >> "$HOME/.zshrc"
+if [ "$DRY_RUN" -ne "1" ]; then
+  # Enable custom config
+  if -f "$HOME/.zshrc"; then
+    if ! grep -q "source $CURRENT_CONFIG_DIR/zsh/config.zsh" "$HOME/.zshrc"; then
+      (echo; echo "source $CURRENT_CONFIG_DIR/zsh/config.zsh") >> "$HOME/.zshrc"
+    fi
+  else
+    echo "source $CURRENT_CONFIG_DIR/zsh/config.zsh" > "$HOME/.zshrc"
   fi
-else
-  echo "source $CURRENT_CONFIG_DIR/zsh/config.zsh" > "$HOME/.zshrc"
 fi
